Added ab encrypter type that will handle all encryption and decryption of packets as well as mac and compression. Currently, its new function handles the key exchange and sends the first packet for ecdh-sha2-nistp256.
Also fixed an off by one error in the ssh stream read function.

Lastly, the exchange hash (which is used in key gen) required so information from initial exchanges so the generation of its prefix (the part including non-key related info) is assembled and passed to the encrypter upon creation.